+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Support for the OpenAI [`gpt-5.2`](https://platform.openai.com/docs/models/gpt-5.2) model.
 - Option to export generated descriptions in TEI XML format, either with line beginning encoding or not. Replaces the previous XML export option.
 - Normalisation of characters in descriptions.
+- Ability to zoom image when editing generated description.
 
 ### Changed